# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API preliminar del Marketplace",
        "description": "Documentación generada con L5-Swagger para la API del Marketplace Orientado a la ciudad de Chivilcoy.",
        "contact": {
            "name": "Santiago Castellano",
            "email": "scastellano10@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/businesses/{business}/discount-tokens": {
            "post": {
                "tags": [
                    "DiscountTokens"
                ],
                "summary": "Generar un token de descuento para un negocio",
                "description": "Genera un token de descuento que el usuario puede presentar al negocio",
                "operationId": "4d2b95e3c3c29b842c7ecd55bfdfb22f",
                "parameters": [
                    {
                        "name": "business",
                        "in": "path",
                        "description": "ID del negocio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "discount_type",
                                    "discount_value",
                                    "valid_days"
                                ],
                                "properties": {
                                    "product_id": {
                                        "description": "ID del producto (opcional)",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "discount_type": {
                                        "description": "Tipo de descuento",
                                        "type": "string",
                                        "enum": [
                                            "percentage",
                                            "fixed"
                                        ],
                                        "example": "percentage"
                                    },
                                    "discount_value": {
                                        "description": "Valor del descuento (10 para 10% o 5.00 para $5)",
                                        "type": "number",
                                        "example": 10
                                    },
                                    "min_purchase": {
                                        "description": "Mínimo de compra requerido (opcional)",
                                        "type": "number",
                                        "example": 50
                                    },
                                    "valid_days": {
                                        "description": "Días de validez desde hoy",
                                        "type": "integer",
                                        "example": 7
                                    },
                                    "description": {
                                        "description": "Descripción del descuento",
                                        "type": "string",
                                        "example": "10% de descuento en tu próxima compra"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Token generado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DiscountToken"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error de validación"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No autorizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/me/discount-tokens": {
            "get": {
                "tags": [
                    "DiscountTokens"
                ],
                "summary": "Listar tokens de descuento del usuario",
                "description": "Devuelve todos los tokens de descuento generados por el usuario actual",
                "operationId": "ebc1d30f3541ece907f61ef92ae6d3a9",
                "responses": {
                    "200": {
                        "description": "Lista de tokens de descuento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DiscountToken"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/discount-tokens/{token}": {
            "get": {
                "tags": [
                    "DiscountTokens"
                ],
                "summary": "Ver detalles de un token de descuento",
                "description": "Devuelve los detalles de un token de descuento usando su código",
                "operationId": "77a66771b139b2d09619525c4ba49de1",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Código del token de descuento",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DiscountToken"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Token no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Token no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/discount-tokens/{token}/use": {
            "post": {
                "tags": [
                    "DiscountTokens"
                ],
                "summary": "Usar un token de descuento",
                "description": "Registra el uso de un token de descuento por parte de un cliente",
                "operationId": "ed6260f9054946a962e74a79abe0f112",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Código del token de descuento",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token usado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Token usado exitosamente"
                                        },
                                        "confirmation_code": {
                                            "description": "Código para que el negocio confirme el uso",
                                            "type": "string",
                                            "example": "ABCD1234"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Token no válido o ya usado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Token no válido o ya usado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Token no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Token no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/discount-tokens/{token}/confirm": {
            "post": {
                "tags": [
                    "DiscountTokens"
                ],
                "summary": "Confirmar uso de un token de descuento",
                "description": "Permite al negocio confirmar que el token fue usado correctamente",
                "operationId": "e8c2048a4384557c188abc11dc4c148a",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Código del token de descuento",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "confirmation_code"
                                ],
                                "properties": {
                                    "confirmation_code": {
                                        "description": "Código de confirmación generado al usar el token",
                                        "type": "string",
                                        "example": "ABCD1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Uso del token confirmado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Uso del token confirmado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Código de confirmación inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Código de confirmación inválido"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado (el usuario no es dueño del negocio)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No autorizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Token no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Token no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DiscountToken": {
                "title": "Token de Descuento",
                "description": "Token de descuento generado por un usuario para un negocio o producto",
                "properties": {
                    "id": {
                        "description": "ID del token",
                        "type": "integer",
                        "example": 1
                    },
                    "code": {
                        "description": "Código único del token",
                        "type": "string",
                        "example": "ABCD-1234"
                    },
                    "business_id": {
                        "description": "ID del negocio",
                        "type": "integer",
                        "example": 21
                    },
                    "product_id": {
                        "description": "ID del producto (nullable)",
                        "type": "integer",
                        "example": 5
                    },
                    "user_id": {
                        "description": "ID del usuario que generó el token",
                        "type": "integer",
                        "example": 3
                    },
                    "discount_type": {
                        "description": "Tipo de descuento",
                        "type": "string",
                        "enum": [
                            "percentage",
                            "fixed"
                        ],
                        "example": "percentage"
                    },
                    "discount_value": {
                        "description": "Valor del descuento (10 para 10% o 5.00 para $5)",
                        "type": "number",
                        "example": 10
                    },
                    "min_purchase": {
                        "description": "Mínimo de compra requerido (nullable)",
                        "type": "number",
                        "example": 50
                    },
                    "max_uses": {
                        "description": "Número máximo de usos",
                        "type": "integer",
                        "example": 1
                    },
                    "uses_count": {
                        "description": "Número de veces que se ha usado",
                        "type": "integer",
                        "example": 0
                    },
                    "valid_from": {
                        "description": "Fecha desde cuando es válido",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-26T00:00:00.000000Z"
                    },
                    "valid_until": {
                        "description": "Fecha hasta cuando es válido",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-02T00:00:00.000000Z"
                    },
                    "description": {
                        "description": "Descripción del descuento",
                        "type": "string",
                        "example": "10% de descuento en tu próxima compra"
                    },
                    "is_active": {
                        "description": "Si el cupón está activo",
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-26T12:34:56.000000Z"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-26T12:34:56.000000Z"
                    },
                    "business": {
                        "$ref": "#/components/schemas/Business"
                    },
                    "product": {
                        "$ref": "#/components/schemas/Product"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}